Add SignupView, remove Authentication components
