+++ apps/web/src/app/auth/signup/page.tsx
@@ -1,14 +1,11 @@
-import Authentication from '../../../components/authentication';
+import SignupView from 'apps/web/src/views/authentication/signup';
 
 import styles from '../page.module.css';
 
 export default function SignupPage() {
   return (
     <div className={styles.wrapper}>
-      <Authentication>
-        <Authentication.Header type="signup" />
-        <Authentication.Form type="signup" />
-      </Authentication>
+      <SignupView />
     </div>
   );
 }
