--- apps/web/src/components/authentication/styles.module.css
@@ -1,8 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-  width: 100%;
-  max-width: 40rem;
-  padding: 3rem;
-}
