+++ apps/web/src/views/authentication/signup/styles.module.css
@@ -1,3 +1,12 @@
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+  width: 100%;
+  max-width: 40rem;
+  padding: 3rem;
+}
+
 .header {
   display: flex;
   flex-direction: column;
